# app/Http/Requests/CustomerCreateRequest.php
<?php

namespace App\Http\Requests;

// use Illuminate\Foundation\Http\FormRequest;
// use Illuminate\Support\Facades\Log;
use App\Http\Requests\ApiFormRequest;

class CustomerCreateRequest extends ApiFormRequest
{
    /**
     * Determine if the user is authorized to make this request.
     */
    public function authorize(): bool
    {
        return true;
    }

    /**
     * Get the validation rules that apply to the request.
     *
     * @return array<string, \Illuminate\Contracts\Validation\ValidationRule|array<mixed>|string>
     */
    public function rules(): array
    {
        return [
            //
            'name' => 'required|string|max:100',
            'address' => 'nullable|string|max:100',
            'phone' => 'nullable|string|max:100',
            'email' => 'required|email|unique:customers|max:120',
        ];
    }
}