# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel",
        "description": "Laravel",
        "contact": {
            "name": "hello@example.com",
            "email": "hello@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local server"
        }
    ],
    "paths": {
        "/api/v1/customers/store": {
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Store new Customer",
                "description": "Store new Customer",
                "operationId": "customer_store",
                "requestBody": {
                    "description": "Register new Customer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerCreateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "400": {
                        "description": "Bad Request"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customers/show/{id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Select existing Customer",
                "description": "Select existing Customer",
                "operationId": "customer_show",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customers/filter": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Filter existing Customers",
                "description": "Filter existing Customers",
                "operationId": "customer_filter",
                "parameters": [
                    {
                        "name": "perPage",
                        "in": "query",
                        "description": "Page Limit",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 10
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "format": "int64",
                            "example": 0
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Search term",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthennticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customers/update/{id}": {
            "put": {
                "tags": [
                    "Customers"
                ],
                "summary": "Update existing Customer",
                "description": "Update existing Customer",
                "operationId": "customer_update",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Update existing Customer",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerUpdateRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/JsonResponseResource"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customers/flag-as-deleted/{id}": {
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Soft Delete existing Customer",
                "description": "Soft Delete existing Customer",
                "operationId": "customer_flagAsDeleted",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
        "/api/v1/customers/permanently-delete-records/{id}": {
            "delete": {
                "tags": [
                    "Customers"
                ],
                "summary": "Force Delete existing Customer",
                "description": "Force Delete existing Customer",
                "operationId": "customer_destroyPermanently",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Customer id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                },
                "security": [
                    {
                        "bearer_token": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CustomerCreateRequest": {
                "title": "Customer Create request",
                "description": "Customer Create request body data",
                "required": [
                    "name"
                ],
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Customer Name",
                        "type": "string",
                        "example": "Kasun Vimarshana"
                    },
                    "address": {
                        "title": "address",
                        "description": "Customer Address",
                        "type": "string",
                        "example": "Address"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Customer Name",
                        "type": "string",
                        "example": "0000000000"
                    },
                    "email": {
                        "title": "email",
                        "description": "Customer Email",
                        "type": "string",
                        "example": "kasunvmail@gmail.com"
                    }
                },
                "type": "object"
            },
...
            "CustomerUpdateRequest": {
                "title": "Customer Update request",
                "description": "Customer Update request body data",
                "properties": {
                    "name": {
                        "title": "name",
                        "description": "Customer Name",
                        "type": "string",
                        "example": "Kasun Vimarshana"
                    },
                    "address": {
                        "title": "address",
                        "description": "Customer Address",
                        "type": "string",
                        "example": "Address"
                    },
                    "phone": {
                        "title": "phone",
                        "description": "Customer Name",
                        "type": "string",
                        "example": "0000000000"
                    },
                    "email": {
                        "title": "email",
                        "description": "Customer Email",
                        "type": "string",
                        "example": "kasunvmail@gmail.com"
                    }
                },
                "type": "object"
            },
...
            "JsonResponseResource": {
                "title": "JsonResponseResource",
                "description": "Json Response Resource",
                "properties": {
                    "status": {
                        "title": "Status",
                        "description": "Status",
                        "type": "string"
                    },
                    "message": {
                        "title": "Message",
                        "description": "Message",
                        "type": "string"
                    },
                    "data": {
                        "title": "Data",
                        "description": "Data",
                        "nullable": true
                    },
                    "errors": {
                        "title": "Errors",
                        "description": "Errors",
                        "nullable": true
                    }
                },
                "type": "object",
                "xml": {
                    "name": "JsonResponseResource"
                }
            },
...
        },
        "securitySchemes": {
            "bearer_token": {
                "type": "http",
                "scheme": "bearer",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}